feat: Create Secure Azure Resources for Linux VM
This commit adds code to create Azure resources for the Secure Azure Resources for Linux VM. It includes the creation of a resource group, VNET, Azure Bastion subnet, VM subnet, Azure Bastion host, and a VM in the VM subnet. The code sets up the necessary variables and uses the Azure CLI commands to create the resources.